data-ingestion-service/requirements.txt
boto3==1.34.34
botocore==1.34.34
psycopg2-binary==2.9.9
mysql-connector-python==8.2.0
pymongo==4.6.0
python-dotenv==1.0.0
pandas==2.1.4
sqlalchemy==2.0.23
flask==3.0.0
flask-cors==4.0.0
schedule==1.2.0
celery==5.3.4
redis==5.0.1
